Move out sjcl util to it's own file to remove injection into runtime injections
